+++ lbryumx/coin.py
@@ -78,6 +78,11 @@ class LBC(Coin):
         )
 
     @classmethod
+    def address_from_script(cls, script):
+        '''Given a pk_script, return the adddress it pays to, or None.'''
+        return ScriptPubKey.pay_to(cls.address_handlers, script)
+
+    @classmethod
     def claim_address_handler(cls, script):
         '''Parse a claim script, returns the address
         '''
